+++ CHANGELOG.md
@@ -21,7 +21,7 @@
   - `bitmap_bit_position`
   - `bitmap_bucket_number`
   - `bitmap_construct_agg`
-  - `bit_shift_right_unsigned`
+  - `bitshiftright_unsigned`
   - `cbrt`
   - `equal_null`
   - `from_json`
@@ -43,7 +43,7 @@
   - `regr_sxx`
   - `regr_sxy`
   - `regr_syy`
-  - `slice`
+  - `array_slice`
   - `try_to_binary`
   - `base64`
   - `base64_decode_string`