+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 ### Snowpark pandas API Updates
 
+#### Dependency Updates
+
+- Added a dependency for `snowflake-ml-python`.
+
 #### New Features
 
 - Added support for applying Snowflake Cortex functions `Summarize` and `Sentiment`.